Updated: - removed try...catch block
