+++ src/main/java/catchmultipleexceptions/InvalidInputParamException.java
@@ -1,4 +1,4 @@
-package catchMultipleExeptions;
+package catchmultipleexceptions;
 
 
 public class InvalidInputParamException extends Exception {
