+++ src/test/java/catchmultipleexceptions/ArithmeticTest.java
@@ -1,4 +1,4 @@
-package catchMultipleExeptions;
+package catchmultipleexceptions;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
